fix: searchParams 관련 이슈 수정
